+++ src/cpp/config.hpp
@@ -17,13 +17,13 @@ class Objective;
 // Training settings
 class Configure {
  public:
-  size_t number_of_feature;      // number of features
-  size_t max_depth;              // max depth for each tree
-  size_t iterations;             // number of trees in gbdt
+  int number_of_feature;      // number of features
+  int max_depth;              // max depth for each tree
+  int iterations;             // number of trees in gbdt
   double shrinkage;               // shrinkage parameter
   double feature_sample_ratio;    // portion of features to be splited
   double data_sample_ratio;       // portion of data to be fitted in each iteration
-  size_t min_leaf_size;          // min number of nodes in leaf
+  int min_leaf_size;          // min number of nodes in leaf
 
   std::shared_ptr<Objective> loss; // loss type
